requirements.txt
# ---- Dependências da Aplicação Core ----
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# ---- Dependências de Teste e Qualidade de Código ----
pytest==8.2.2
pylint==3.2.5